requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
fonttools==4.34.4
idna==3.3
joblib==1.1.0
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.2
multitasking==0.0.11
mysql==0.0.3
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.23.1
packaging==21.3
pandas==1.4.3
patsy==0.5.2
Pillow==9.2.0
protobuf==3.20.1
psutil==5.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
threadpoolctl==3.1.0
urllib3==1.26.11
websockets==10.3
yfinance==0.1.74
